fix(routes): normalize remaining repeated-segment pages
